Fix for mgcv::gam objects without smooth terms
The new code for mgcv::gam objects looks great, but unfortunately breaks down when there is no smooth term in the model. This is due to object$smooth being an empty list, i.e. list(), instead of NULL when there is no smooth term; so I've implemented the checks via length(object$smooth) > 0.
I've added two more suggestions: It might be better to ensure that NULL instead of logical(0) is passed to the exclude argument of mgcv::predict.gam when there is no random term, and one might want to ensure that all columns are kept when there is no smooth term or random term (there might exist 0-columns in X for some other reason).